Create gfgpotd_111223.java
It was our 3rd day in Gowardhan Dandwati paricharma. It was full of excitement and Fun. We had breakfast and Tea and Lunch on the same paricharma marg. I was amazed when I meet an old dadi who was also doing Dandwati paricharma. I found a maharaj who does paricharma twice a day. Today we crossed Uddhav kund. We are 10 people doing paricharma. And 2+3 people are doing seva. We Also visited mansi sarowar bazaar. Tomorrow hopefully we will cross Radha kund. When I achieve a target specified I try to help everyone to motivate and fun. We also dance sometimes when needed.